Silenced pedantic compiler warnings

+++ gmime/gmime-multipart-signed.c
@@ -348,7 +348,7 @@ check_protocol_supported (const char *protocol, const char *supported)
 	 * protocol. For example, if the supported protocol is
 	 * "application/pkcs7-signature", then we also want to
 	 * match "application/x-pkcs7-signature". */
-	xsupported = g_strdup_printf ("%.*sx-%s", subtype - supported, supported, subtype);
+	xsupported = g_strdup_printf ("%.*sx-%s", (int) (subtype - supported), supported, subtype);
 	rv = !g_ascii_strcasecmp (protocol, xsupported);
 	g_free (xsupported);
